[FEATURE] Team 엔티티 생성
