+++ src/main/java/com/sports/server/command/leagueteam/domain/Unit.java
@@ -1,0 +1,8 @@
+package com.sports.server.command.leagueteam.domain;
+
+public enum Unit {
+    HUFS_1,
+    HUFS_2,
+    HUFS_3;
+
+}
