--- src/main/java/com/sports/server/command/team/domain/TeamUnit.java
@@ -1,8 +1,0 @@
-package com.sports.server.command.team.domain;
-
-public enum TeamUnit {
-    HUFS_1,
-    HUFS_2,
-    HUFS_3;
-
-}
