--- src/main/java/com/sports/server/query/dto/response/LeagueSportResponse.java
@@ -1,13 +1,0 @@
-package com.sports.server.query.dto.response;
-
-public record LeagueSportResponse(
-        Long sportId,
-        String name
-) {
-    public LeagueSportResponse(LeagueSport leagueSport) {
-        this(
-                leagueSport.getSport().getId(),
-                leagueSport.getSport().getName()
-        );
-    }
-}
