--- src/main/java/com/sports/server/query/dto/response/SportResponse.java
@@ -1,14 +1,0 @@
-package com.sports.server.query.dto.response;
-
-public record SportResponse(
-        Long id,
-        String name
-
-) {
-    public SportResponse(Sport sport) {
-        this(
-                sport.getId(),
-                sport.getName()
-        );
-    }
-}
