+++ src/test/java/com/sports/server/query/acceptance/SportQueryAcceptanceTest.java
@@ -3,7 +3,6 @@ package com.sports.server.query.acceptance;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import com.sports.server.query.dto.response.SportResponse;
 import com.sports.server.support.AcceptanceTest;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
